# src/main/java/atlassports/api/SubscriptionsApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.18.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package atlassports.api;

import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.validation.constraints.NotNull;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.context.request.NativeWebRequest;

import javax.annotation.processing.Generated;
import java.util.List;
import java.util.Optional;
import java.util.UUID;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-12-24T16:37:25.216838800+01:00[Europe/Skopje]", comments = "Generator version: 7.18.0")
@Validated
@Tag(name = "Subscriptions", description = "Subscription management")
public interface SubscriptionsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_SUBSCRIPTIONS_GET = "/subscriptions";

    /**
     * GET /subscriptions : Get all subscriptions
     *
     * @return List of subscriptions (status code 200)
     */
    @Operation(
            operationId = "subscriptionsGet",
            summary = "Get all subscriptions",
            tags = {"Subscriptions"},
            responses = {
                    @ApiResponse(responseCode = "200", description = "List of subscriptions", content = {
                            @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Object.class)))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.GET,
            value = SubscriptionsApi.PATH_SUBSCRIPTIONS_GET,
            produces = {"application/json"}
    )
    default ResponseEntity<List<Object>> subscriptionsGet(

    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType : MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ null, null ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_SUBSCRIPTIONS_SUBSCRIPTION_ID_GET = "/subscriptions/{subscriptionId}";

    /**
     * GET /subscriptions/{subscriptionId} : Get subscriptions details
     *
     * @param subscriptionId (required)
     * @return Subscription details (status code 200)
     */
    @Operation(
            operationId = "subscriptionsSubscriptionIdGet",
            summary = "Get subscriptions details",
            tags = {"Subscriptions"},
            responses = {
                    @ApiResponse(responseCode = "200", description = "Subscription details", content = {
                            @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Object.class)))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.GET,
            value = SubscriptionsApi.PATH_SUBSCRIPTIONS_SUBSCRIPTION_ID_GET,
            produces = {"application/json"}
    )
    default ResponseEntity<List<Object>> subscriptionsSubscriptionIdGet(
            @NotNull @Parameter(name = "subscriptionId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("subscriptionId") UUID subscriptionId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType : MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ null, null ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_SUBSCRIPTIONS_TENANTS_TENANT_ID_GET = "/subscriptions/tenants/{tenantId}";

    /**
     * GET /subscriptions/tenants/{tenantId} : Get all subscriptions for tenant
     *
     * @param tenantId (required)
     * @return List of subscriptions for tenant (status code 200)
     */
    @Operation(
            operationId = "subscriptionsTenantsTenantIdGet",
            summary = "Get all subscriptions for tenant",
            tags = {"Subscriptions"},
            responses = {
                    @ApiResponse(responseCode = "200", description = "List of subscriptions for tenant", content = {
                            @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Object.class)))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.GET,
            value = SubscriptionsApi.PATH_SUBSCRIPTIONS_TENANTS_TENANT_ID_GET,
            produces = {"application/json"}
    )
    default ResponseEntity<List<Object>> subscriptionsTenantsTenantIdGet(
            @NotNull @Parameter(name = "tenantId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("tenantId") UUID tenantId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType : MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ null, null ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}