# src/main/java/atlassports/api/TenantsApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.18.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package atlassports.api;

import atlassports.model.openapi.TenantsGet200Response;
import atlassports.model.openapi.TenantsTenantIdStatusPatchRequest;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.validation.Valid;
import jakarta.validation.constraints.Max;
import jakarta.validation.constraints.Min;
import jakarta.validation.constraints.NotNull;
import org.springframework.format.annotation.DateTimeFormat;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.lang.Nullable;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;

import javax.annotation.processing.Generated;
import java.time.LocalDate;
import java.util.Optional;
import java.util.UUID;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-12-24T16:37:25.216838800+01:00[Europe/Skopje]", comments = "Generator version: 7.18.0")
@Validated
@Tag(name = "Tenants", description = "Tenant management")
public interface TenantsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_TENANTS_GET = "/tenants";

    /**
     * GET /tenants : Get all tenants (Admin only)
     *
     * @param page     (optional, default to 0)
     * @param size     (optional, default to 20)
     * @param sort     (optional)
     * @param dateFrom Filter tenants created from this date (ISO 8601 format) (optional)
     * @param dateTo   Filter tenants created up to this date (ISO 8601 format) (optional)
     * @return List of tenants (status code 200)
     */
    @Operation(
            operationId = "tenantsGet",
            summary = "Get all tenants (Admin only)",
            tags = {"Tenants"},
            responses = {
                    @ApiResponse(responseCode = "200", description = "List of tenants", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = TenantsGet200Response.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.GET,
            value = TenantsApi.PATH_TENANTS_GET,
            produces = {"application/json"}
    )
    default ResponseEntity<TenantsGet200Response> tenantsGet(
            @Min(value = 0) @Parameter(name = "page", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "0") Integer page,
            @Min(value = 1) @Max(value = 100) @Parameter(name = "size", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false, defaultValue = "20") Integer size,
            @Parameter(name = "sort", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "sort", required = false) @Nullable String sort,
            @Parameter(name = "dateFrom", description = "Filter tenants created from this date (ISO 8601 format)", in = ParameterIn.QUERY) @Valid @RequestParam(value = "dateFrom", required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) @Nullable LocalDate dateFrom,
            @Parameter(name = "dateTo", description = "Filter tenants created up to this date (ISO 8601 format)", in = ParameterIn.QUERY) @Valid @RequestParam(value = "dateTo", required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) @Nullable LocalDate dateTo
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType : MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"size\" : 6, \"last\" : true, \"totalPages\" : 5, \"page\" : 0, \"content\" : [ null, null ], \"first\" : true, \"totalElements\" : 1 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_TENANTS_POST = "/tenants";

    /**
     * POST /tenants : Create new tenant (Admin only)
     *
     * @param body (required)
     * @return Tenant created (status code 201)
     */
    @Operation(
            operationId = "tenantsPost",
            summary = "Create new tenant (Admin only)",
            tags = {"Tenants"},
            responses = {
                    @ApiResponse(responseCode = "201", description = "Tenant created", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = Object.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.POST,
            value = TenantsApi.PATH_TENANTS_POST,
            produces = {"application/json"},
            consumes = {"application/json"}
    )
    default ResponseEntity<Object> tenantsPost(
            @Parameter(name = "body", description = "", required = true) @Valid @RequestBody Object body
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_TENANTS_TENANT_ID_GET = "/tenants/{tenantId}";

    /**
     * GET /tenants/{tenantId} : Get tenant details
     *
     * @param tenantId (required)
     * @return Tenant details (status code 200)
     */
    @Operation(
            operationId = "tenantsTenantIdGet",
            summary = "Get tenant details",
            tags = {"Tenants"},
            responses = {
                    @ApiResponse(responseCode = "200", description = "Tenant details", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = Object.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.GET,
            value = TenantsApi.PATH_TENANTS_TENANT_ID_GET,
            produces = {"application/json"}
    )
    default ResponseEntity<Object> tenantsTenantIdGet(
            @NotNull @Parameter(name = "tenantId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("tenantId") UUID tenantId
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_TENANTS_TENANT_ID_PATCH = "/tenants/{tenantId}";

    /**
     * PATCH /tenants/{tenantId} : Update tenant
     *
     * @param tenantId (required)
     * @param body     (required)
     * @return Updated tenant (status code 200)
     */
    @Operation(
            operationId = "tenantsTenantIdPatch",
            summary = "Update tenant",
            tags = {"Tenants"},
            responses = {
                    @ApiResponse(responseCode = "200", description = "Updated tenant", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = Object.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.PATCH,
            value = TenantsApi.PATH_TENANTS_TENANT_ID_PATCH,
            produces = {"application/json"},
            consumes = {"application/json"}
    )
    default ResponseEntity<Object> tenantsTenantIdPatch(
            @NotNull @Parameter(name = "tenantId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("tenantId") UUID tenantId,
            @Parameter(name = "body", description = "", required = true) @Valid @RequestBody Object body
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_TENANTS_TENANT_ID_STATUS_PATCH = "/tenants/{tenantId}/status";

    /**
     * PATCH /tenants/{tenantId}/status : Enable or disable tenant
     *
     * @param tenantId                          (required)
     * @param tenantsTenantIdStatusPatchRequest (required)
     * @return Tenant status updated (status code 200)
     */
    @Operation(
            operationId = "tenantsTenantIdStatusPatch",
            summary = "Enable or disable tenant",
            tags = {"Tenants"},
            responses = {
                    @ApiResponse(responseCode = "200", description = "Tenant status updated")
            }
    )
    @RequestMapping(
            method = RequestMethod.PATCH,
            value = TenantsApi.PATH_TENANTS_TENANT_ID_STATUS_PATCH,
            consumes = {"application/json"}
    )
    default ResponseEntity<Void> tenantsTenantIdStatusPatch(
            @NotNull @Parameter(name = "tenantId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("tenantId") UUID tenantId,
            @Parameter(name = "TenantsTenantIdStatusPatchRequest", description = "", required = true) @Valid @RequestBody TenantsTenantIdStatusPatchRequest tenantsTenantIdStatusPatchRequest
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}