# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Atlas Sports API
  version: 1.0.0
servers:
- url: http://localhost:8080/api
...
paths:
...
  /tenants:
    get:
      parameters:
      - explode: true
        in: query
        name: page
        required: false
        schema:
          default: 0
          minimum: 0
          type: integer
        style: form
      - explode: true
        in: query
        name: size
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          type: integer
        style: form
      - explode: true
        in: query
        name: sort
        required: false
        schema:
          example: "createdAt,desc"
          type: string
        style: form
      - description: Filter tenants created from this date (ISO 8601 format)
        explode: true
        in: query
        name: dateFrom
        required: false
        schema:
          format: date
          type: string
        style: form
      - description: Filter tenants created up to this date (ISO 8601 format)
        explode: true
        in: query
        name: dateTo
        required: false
        schema:
          format: date
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_tenants_get_200_response"
          description: List of tenants
      summary: Get all tenants (Admin only)
      tags:
      - Tenants
      x-accepts:
      - application/json
      x-tags:
      - tag: Tenants
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertTenantDto"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantDto"
          description: Tenant created
      summary: Create new tenant (Admin only)
      tags:
      - Tenants
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: Tenants
  /tenants/{tenantId}:
    get:
      parameters:
      - explode: false
        in: path
        name: tenantId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantDto"
          description: Tenant details
      summary: Get tenant details
      tags:
      - Tenants
      x-accepts:
      - application/json
      x-tags:
      - tag: Tenants
    patch:
      parameters:
      - explode: false
        in: path
        name: tenantId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertTenantDto"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantDto"
          description: Updated tenant
      summary: Update tenant
      tags:
      - Tenants
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: Tenants
  /tenants/{tenantId}/status:
    patch:
      parameters:
      - explode: false
        in: path
        name: tenantId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/_tenants__tenantId__status_patch_request"
        required: true
      responses:
        "200":
          description: Tenant status updated
      summary: Enable or disable tenant
      tags:
      - Tenants
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: Tenants
...
  /subscriptions/tenants/{tenantId}:
    get:
      parameters:
      - explode: false
        in: path
        name: tenantId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/SubscriptionDto"
                type: array
          description: List of subscriptions for tenant
      summary: Get all subscriptions for tenant
      tags:
      - Subscriptions
      x-accepts:
      - application/json
      x-tags:
      - tag: Subscriptions
  /subscriptions/{subscriptionId}:
    get:
      parameters:
      - explode: false
        in: path
        name: subscriptionId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/SubscriptionDto"
                type: array
          description: Subscription details
      summary: Get subscriptions details
      tags:
      - Subscriptions
      x-accepts:
      - application/json
      x-tags:
      - tag: Subscriptions
...
components:
...
  schemas:
...
    TenantDto:
      type: object
    UpsertTenantDto:
      type: object
...
    SubscriptionDto:
      type: object
...
    PageResponse:
      properties:
        content:
          items:
            type: object
          type: array
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
      type: object
    _tenants_get_200_response:
      allOf:
      - $ref: "#/components/schemas/PageResponse"
      - properties:
          content:
            items:
              $ref: "#/components/schemas/TenantDto"
            type: array
        type: object
      example:
        size: 6
        last: true
        totalPages: 5
        page: 0
        content:
        - null
        - null
        first: true
        totalElements: 1
    _tenants__tenantId__status_patch_request:
      properties:
        active:
          type: boolean
      type: object
...